requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.1
django-environ==0.11.2
django-markdownx==4.0.7
idna==3.6
Markdown==3.7
pillow==10.4.0
psycopg2==2.9.9
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
wheel==0.42.0
